Added RosterManager
