+++ voice-client-example/src/main/java/com/tuenti/voice/example/service/IRosterService.aidl
@@ -1,0 +1,10 @@
+package com.tuenti.voice.example.service;
+
+import com.tuenti.voice.example.service.IRosterServiceCallback;
+
+interface IRosterService {
+
+    void registerCallback( IRosterServiceCallback cb );
+    void unregisterCallback( IRosterServiceCallback cb );
+
+}
