+++ voice-client-example/src/main/java/com/tuenti/voice/example/service/IRosterServiceCallback.aidl
@@ -1,0 +1,4 @@
+package com.tuenti.voice.example.service;
+
+oneway interface IRosterServiceCallback {
+}
