+++ voice-client-example/src/main/java/com/tuenti/voice/example/service/VoiceClientService.java
@@ -3,19 +3,13 @@ package com.tuenti.voice.example.service;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.util.Log;
-import com.tuenti.voice.core.BuddyListState;
-import com.tuenti.voice.core.RosterListener;
 import com.tuenti.voice.core.VoiceClient;
 import com.tuenti.voice.example.util.ConnectionMonitor;
 
 public class VoiceClientService
     extends Service
-    implements RosterListener
 {
 // ------------------------------ FIELDS ------------------------------
-
-    private static final String TAG = "VoiceClientService";
 
     private CallManager mCallManager;
 
@@ -23,32 +17,7 @@ public class VoiceClientService
 
     private ConnectionManager mConnectionManager;
 
-// ------------------------ INTERFACE METHODS ------------------------
-
-// --------------------- Interface RosterListener ---------------------
-
-    @Override
-    public void handleBuddyListChanged( int state, String remoteJid )
-    {
-        switch ( BuddyListState.fromInteger( state ) )
-        {
-            case ADD:
-                Log.v( TAG, "Adding buddy " + remoteJid );
-                // Intent add buddy
-                // mBuddyList.add(remoteJid);
-                break;
-            case REMOVE:
-                Log.v( TAG, "Removing buddy" + remoteJid );
-                // Intent remove buddy
-                // mBuddyList.remove(remoteJid);
-                break;
-            case RESET:
-                Log.v( TAG, "Reset buddy list" );
-                // intent reset buddy list
-                // mBuddyList.clear();
-                break;
-        }
-    }
+    private RosterManager mRosterManager;
 
 // -------------------------- OTHER METHODS --------------------------
 
@@ -58,6 +27,10 @@ public class VoiceClientService
         if ( IConnectionService.class.getName().equals( intent.getAction() ) )
         {
             return mConnectionManager.onBind();
+        }
+        if ( IRosterService.class.getName().equals( intent.getAction() ) )
+        {
+            return mRosterManager.onBind();
         }
         if ( ICallService.class.getName().equals( intent.getAction() ) )
         {
@@ -71,10 +44,12 @@ public class VoiceClientService
     {
         super.onCreate();
 
+        // VoiceClient should only be created here
+        // probably init here too.
         mClient = new VoiceClient();
-        mClient.addRosterListener( this );
 
         mConnectionManager = new ConnectionManager( mClient );
+        mRosterManager = new RosterManager( mClient );
         mCallManager = new CallManager( mClient, getBaseContext() );
 
         ConnectionMonitor.getInstance( getApplicationContext() );
